fix: 루트 경로(/) 로그인 페이지로 변경, 라다이렉트 수정
