+++ src/lib/components/ui/sidebar.svelte
@@ -91,7 +91,7 @@
             removeToken();
 
             // 로그인 페이지로 리다이렉트
-            goto('/login');
+            goto('/');
         }
     }
 </script>
