+++ src/routes/+layout.ts
@@ -1,13 +1,14 @@
-export const load = ({ url }) => {
-  const path = url.pathname;
+export const load = ({url}) => {
+    const path = url.pathname;
 
-  const pathTitleMap: Record<string, string> = {
-    '/': 'Main',
-    '/place': 'Place Management',
-    '/projects': 'Project Management',
-  };
+    const pathTitleMap: Record<string, string> = {
+        '/': 'Login',
+        '/home': 'Home',
+        '/place': 'Place Management',
+        '/projects': 'Project Management',
+    };
 
-  return {
-    pageTitle: pathTitleMap[path] ?? '없는 페이지',
-  };
+    return {
+        pageTitle: pathTitleMap[path] ?? '없는 페이지',
+    };
 };
