+++ src/routes/+page.svelte
@@ -1,8 +1,5 @@
 <script lang="ts">
-import { HomePage } from '$lib/pages';
-import type { PageData, ActionData } from './$types';
-
-let { data, form }: { data: PageData, form?: ActionData } = $props();
+    import { LoginPage } from '$lib/pages';
 </script>
 
-<HomePage {data} {form} />
+<LoginPage />
