--- src/routes/login/+page.svelte
@@ -1,5 +1,0 @@
-<script lang="ts">
-    import { LoginPage } from '$lib/pages';
-</script>
-
-<LoginPage />
